requirements.txt
opencv-python~=4.10.0.84
onnxruntime~=1.19.2
numpy~=2.0.2
torch~=2.4.1
flask~=3.0.3
pillow~=10.4.0
watchdog~=5.0.3
six~=1.16.0